+++ tests/transformers_utils/test_tokenizer_registry.py
@@ -1,8 +1,7 @@
 # SPDX-License-Identifier: Apache-2.0
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
-from vllm.transformers_utils.registry import TokenizerRegistry
 from vllm.transformers_utils.tokenizer import get_tokenizer
-from vllm.transformers_utils.tokenizers import TokenizerLike
+from vllm.transformers_utils.tokenizers import TokenizerLike, TokenizerRegistry
 
 
 class TestTokenizer(TokenizerLike):
