+++ vllm/transformers_utils/tokenizer_base.py
@@ -9,8 +9,8 @@ def __getattr__(name: str):
         from .tokenizers import TokenizerLike
 
         warnings.warn(
-            "`vllm.transformers_utils.tokenizer_base.TokenizerBase` "
-            "has been moved to `vllm.transformers_utils.tokenizers.TokenizerLike`. "
+            "`vllm.transformers_utils.tokenizer_base.TokenizerBase` has been "
+            "moved to `vllm.transformers_utils.tokenizers.TokenizerLike`. "
             "The old name will be removed in v0.13.",
             DeprecationWarning,
             stacklevel=2,
@@ -18,11 +18,11 @@ def __getattr__(name: str):
 
         return TokenizerLike
     if name == "TokenizerRegistry":
-        from .registry import TokenizerRegistry
+        from .tokenizers import TokenizerRegistry
 
         warnings.warn(
-            "`vllm.transformers_utils.tokenizer_base.TokenizerRegistry` "
-            "has been moved to `vllm.transformers_utils.registry.TokenizerRegistry`. "
+            "`vllm.transformers_utils.tokenizer_base.TokenizerRegistry` has been "
+            "moved to `vllm.transformers_utils.tokenizers.TokenizerRegistry`. "
             "The old name will be removed in v0.13.",
             DeprecationWarning,
             stacklevel=2,
