+++ vllm/transformers_utils/tokenizers/__init__.py
@@ -1,18 +1,8 @@
 # SPDX-License-Identifier: Apache-2.0
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 
-from .mistral import (
-    MistralTokenizer,
-    maybe_serialize_tool_calls,
-    truncate_tool_call_ids,
-    validate_request_params,
-)
+from .mistral import MistralTokenizer
 from .protocol import TokenizerLike
+from .registry import TokenizerRegistry
 
-__all__ = [
-    "TokenizerLike",
-    "MistralTokenizer",
-    "maybe_serialize_tool_calls",
-    "truncate_tool_call_ids",
-    "validate_request_params",
-]
+__all__ = ["TokenizerLike", "MistralTokenizer", "TokenizerRegistry"]
